Insert cettification added
Currently everything in insert_pp workds which is the code for a student to submit course enrollment and certification enrollment. For the form pages for this still have to go back and add descriptions on ways to insert the data for example Prog_Num should show which numbers correspond to what programs. But as far as functionality the pages correctly insert into the database a new course or new cert